+++ core/src/app_core.rs
@@ -272,6 +272,30 @@ impl AppCore {
             })?;
         Ok(())
     }
+
+    /// Change the address to connect with to a dapp.
+    /// Assumes there is already a key for the dapp in the account.
+    pub fn eth_change_dapp_chain(
+        &self,
+        args: EthChangeDappChainArgs,
+    ) -> Result<(), CoreError> {
+        let new_chain_id: eth::ChainId = args.new_chain_id.try_into()?;
+        self.connection_pool()
+            .deferred_transaction(move |mut tx_conn| {
+                let params = m::NewDappSessionParams::builder()
+                    .account_id(&args.account_id)
+                    .dapp_id(&args.dapp_id)
+                    .chain_id(new_chain_id)
+                    .build();
+                let session = m::LocalDappSession::create_eth_session_if_not_exists(
+                    &mut tx_conn,
+                    &params,
+                )?;
+                session.change_eth_chain(&mut tx_conn, new_chain_id)?;
+                Ok(())
+            })?;
+        Ok(())
+    }
 }
 
 #[derive(Debug)]
@@ -293,6 +317,13 @@ pub struct EthTransferFungibleTokenArgs {
     pub to_checksum_address: String,
     pub amount_decimal: String,
     pub token_id: String,
+}
+
+#[derive(Debug, Clone, TypedBuilder)]
+pub struct EthChangeDappChainArgs {
+    pub account_id: String,
+    pub dapp_id: String,
+    pub new_chain_id: u64,
 }
 
 #[derive(Debug, Clone)]